bot/wsgi.py
"""
WSGI config for bot project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/3.0/howto/deployment/wsgi/
"""

import os
from django.core.wsgi import get_wsgi_application
from dotenv import load_dotenv

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'bot.settings')

application = get_wsgi_application()

chat = os.path.expanduser('~/bot')
load_dotenv(os.path.join(chat, '.env'))
