requirements.txt
asgiref==3.2.10
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.2
chardet==3.0.4
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==7.1.2
cymem==2.0.3
Django==3.0.7
future==0.18.2
idna==2.9
joblib==0.15.1
mathparse==0.1.2
murmurhash==1.0.2
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
packaging==20.4
Pint==0.13
plac==1.1.3
preshed==3.0.2
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.7.5
python-dotenv==0.13.0
python-twitter==3.5
pytz==2020.1
PyYAML==3.13
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
spacy==2.2.3
SQLAlchemy==1.2.19
sqlparse==0.3.1
srsly==1.0.2
thinc==7.3.1
tqdm==4.46.1
urllib3==1.25.9
wasabi==0.6.0
